Webpack aliases, tslint rules, packages upgrade
