+++ package.json
@@ -45,7 +45,7 @@
   },
   "dependencies": {
     "babel-loader": "^8.0.4",
-    "cache-loader": "^1.2.5",
+    "cache-loader": "^2.0.0",
     "clean-webpack-plugin": "^1.0.0",
     "cross-env": "^5.2.0",
     "css-loader": "^1.0.1",
@@ -61,7 +61,7 @@
     "modernizr-loader": "^1.0.1",
     "optimize-css-assets-webpack-plugin": "^5.0.1",
     "path": "^0.12.7",
-    "preload-webpack-plugin": "^3.0.0-beta.1",
+    "preload-webpack-plugin": "^3.0.0-beta.3",
     "script-ext-html-webpack-plugin": "^2.1.3",
     "style-loader": "^0.23.1",
     "terser-webpack-plugin": "^1.2.1",
@@ -72,22 +72,22 @@
     "workbox-webpack-plugin": "^3.6.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.1.6",
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
-    "@babel/plugin-proposal-decorators": "^7.1.6",
-    "@babel/plugin-proposal-export-default-from": "^7.0.0",
-    "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
-    "@babel/plugin-proposal-function-bind": "^7.0.0",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.0.0",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.0.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
-    "@babel/plugin-proposal-optional-chaining": "^7.0.0",
-    "@babel/plugin-proposal-pipeline-operator": "^7.0.0",
-    "@babel/plugin-proposal-throw-expressions": "^7.0.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
-    "@babel/plugin-transform-react-constant-elements": "^7.0.0",
-    "@babel/plugin-transform-react-inline-elements": "^7.0.0",
-    "@babel/preset-env": "^7.1.6",
+    "@babel/core": "^7.2.2",
+    "@babel/plugin-proposal-class-properties": "^7.2.3",
+    "@babel/plugin-proposal-decorators": "^7.2.3",
+    "@babel/plugin-proposal-export-default-from": "^7.2.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
+    "@babel/plugin-proposal-function-bind": "^7.2.0",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.2.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.2.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.2.0",
+    "@babel/plugin-proposal-optional-chaining": "^7.2.0",
+    "@babel/plugin-proposal-pipeline-operator": "^7.2.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.2.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/plugin-transform-react-constant-elements": "^7.2.0",
+    "@babel/plugin-transform-react-inline-elements": "^7.2.0",
+    "@babel/preset-env": "^7.2.3",
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.1.0",
     "@storybook/addon-actions": "^4.0.8",
@@ -96,7 +96,7 @@
     "@storybook/addon-viewport": "^4.0.8",
     "@storybook/addons": "^4.0.8",
     "@storybook/react": "^4.0.8",
-    "autoprefixer": "^9.3.1",
+    "autoprefixer": "^9.4.3",
     "babel-core": "^7.0.0-0",
     "babel-jest": "^23.6.0",
     "babel-plugin-closure-elimination": "^1.3.0",
@@ -105,15 +105,15 @@
     "babel-plugin-module-resolver": "^3.1.1",
     "babel-plugin-react-css-modules": "^5.0.0",
     "css-mqpacker": "^7.0.0",
-    "cssnano": "^4.1.7",
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.7.0",
-    "enzyme-to-json": "^3.3.4",
+    "cssnano": "^4.1.8",
+    "enzyme": "^3.8.0",
+    "enzyme-adapter-react-16": "^1.7.1",
+    "enzyme-to-json": "^3.3.5",
     "fork-ts-checker-webpack-plugin": "^0.5.2",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^23.6.0",
     "jest-cli": "^23.6.0",
-    "postcss": "^7.0.6",
+    "postcss": "^7.0.7",
     "postcss-advanced-variables": "^3.0.0",
     "postcss-aspect-ratio": "^1.0.2",
     "postcss-color-function": "^4.0.1",
@@ -122,15 +122,15 @@
     "postcss-import": "^12.0.1",
     "postcss-line-height-px-to-unitless": "^0.2.1",
     "postcss-loader": "^3.0.0",
-    "postcss-nested": "^4.1.0",
-    "postcss-preset-env": "^6.4.0",
+    "postcss-nested": "^4.1.1",
+    "postcss-preset-env": "^6.5.0",
     "postcss-pxtorem": "^4.0.1",
     "postcss-scale": "^1.0.1",
     "postcss-scss": "^2.0.0",
     "postcss-selector-matches": "^4.0.0",
     "postcss-selector-not": "^4.0.0",
     "postcss-svg": "^3.0.0",
-    "react-test-renderer": "^16.6.3",
+    "react-test-renderer": "^16.7.0",
     "redux-mock-store": "^1.5.3",
     "storybook-readme": "^4.0.2",
     "typescript": "^3.2.2"
@@ -146,8 +146,8 @@
     "eslint-config-standard-react": "^7.0.2",
     "eslint-import-resolver-webpack": "^0.10.1",
     "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jest": "^22.0.0",
-    "eslint-plugin-json": "^1.2.1",
+    "eslint-plugin-jest": "^22.1.2",
+    "eslint-plugin-json": "^1.3.2",
     "eslint-plugin-jsx-a11y": "^6.1.2",
     "eslint-plugin-node": "^8.0.0",
     "eslint-plugin-prettier": "^3.0.1",
@@ -158,7 +158,7 @@
     "lint-staged": "^8.1.0",
     "onchange": "^5.2.0",
     "prettier": "^1.15.3",
-    "stylelint": "^9.8.0",
+    "stylelint": "^9.9.0",
     "stylelint-at-rule-no-children": "^0.3.1",
     "stylelint-color-format": "^0.2.0",
     "stylelint-config-airbnb": "^0.0.0",
