+++ src/js/components/app.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import { hot } from 'react-hot-loader/root'
+console.log('Find me!')
+
+function App() {
+  return <button>Hello, TS =)</button>
+}
+
+export default hot(App)
