+++ src/package.json
@@ -12,13 +12,13 @@
     "babel-plugin-react-css-modules": "^5.0.0",
     "connected-react-router": "^6.1.0",
     "history": "^4.7.2",
-    "hoist-non-react-statics": "^3.1.0",
-    "immer": "^1.8.0",
+    "hoist-non-react-statics": "^3.2.1",
+    "immer": "^1.9.3",
     "lodash-es": "^4.17.11",
-    "react": "^16.6.3",
+    "react": "^16.7.0",
     "react-cached-callback": "^1.0.1",
     "react-compose-context-consumers": "^0.1.3",
-    "react-dom": "^16.6.3",
+    "react-dom": "^16.7.0",
     "react-loadable": "^5.5.0",
     "react-redux": "^6.0.0",
     "react-router-dom": "^4.3.1",
@@ -26,8 +26,9 @@
     "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
-    "react-hot-loader": "^4.3.12",
+    "react-hot-loader": "^4.6.3",
     "@types/react": "^16.7.18",
-    "@types/react-dom": "^16.0.11"
+    "@types/react-dom": "^16.0.11",
+    "@types/react-redux": "^6.0.11"
   }
 }
