+++ tsconfig.json
@@ -7,12 +7,23 @@
     "moduleResolution": "node",
     "esModuleInterop": true,
     "noUnusedLocals": true,
-    "noImplicitAny": true,
+    "noImplicitAny": false,
     "target": "es2015",
-    "module": "es2015",
+    "module": "esnext",
     "strict": true,
-    "jsx": "react"
+    "jsx": "react",
+    "baseUrl": "./",
+    "paths": {
+      "~components/*": ["src/js/components/*"],
+      "~store/*": ["src/js/store/*"],
+      "~utils/*": ["src/js/utils/*"],
+      "~constants": ["src/js/constants.js"],
+      "~modules/*": ["src/js/modules/*"],
+      "~css/*": ["src/css/*"],
+      "~img/*": ["src/img/*"],
+      "modernizr$": [".modernizrrc.js"]
+    }
   },
   "include": ["src/js/**/*"],
-  "exclude": ["/**/node_modules", "dist"]
+  "exclude": ["/**/node_modules"]
 }
