+++ tslint.json
@@ -1,21 +1,24 @@
 {
-    "defaultSeverity": "error",
-    "extends": [
-        "tslint:recommended",
-        "tslint-react",
-        "tslint-config-airbnb",
-        "tslint-sonarts",
-        "tslint-immutable",
-        "tslint-config-prettier"
-    ],
-    "jsRules": {
-
-    },
-    "rules": {
-        "prettier": true,
-        "no-constant-condition": true,
-        "import-name": false,
-        "function-name": false
-    },
-    "rulesDirectory": ["tslint-plugin-prettier"]
+  "defaultSeverity": "error",
+  "extends": [
+    "tslint:recommended",
+    "tslint-react",
+    "tslint-config-airbnb",
+    "tslint-sonarts",
+    "tslint-immutable",
+    "tslint-config-prettier"
+  ],
+  "jsRules": {},
+  "rules": {
+    "prettier": true,
+    "no-constant-condition": true,
+    "import-name": false,
+    "function-name": false,
+    "no-console": false,
+    "ordered-imports": false,
+    "no-commented-code": false,
+    "one-variable-per-declaration": false,
+    "variable-name": false
+  },
+  "rulesDirectory": ["tslint-plugin-prettier"]
 }
